Restructured code and better collision system
Code restructured - instead creating a for loop to create clickable points another for stones, etc, do it all in one for loop.
Collision system improved - the source of the low frame rate was the board collisions. So split up the board and only test for collisions on the part of the board that it is near.